Add integration test for AGP 9.0
